Added working TorchStreamWriter
