+++ src/wsi_patching/writers/__init__.py
@@ -1,6 +1,6 @@
-from .numpy_mem_writer import NumpyMemoryWriter
-from .torch_mem_writer import TorchMemoryWriter
+from .numpy_stream_writer import NumpyStreamWriter
+from .torch_stream_writer import TorchStreamWriter
 from .webdataset.webdataset_loader import WebDatasetLoader
 from .webdataset.webdataset_writer import WebDatasetWriter
 
-__all__ = ["NumpyMemoryWriter", "TorchMemoryWriter", "WebDatasetWriter", "WebDatasetLoader"]
+__all__ = ["NumpyStreamWriter", "TorchStreamWriter", "WebDatasetWriter", "WebDatasetLoader"]
