+++ src/wsi_patching/writers/numpy_stream_writer.py
@@ -7,24 +7,15 @@ from wsi_patching.core.types.types import CollatedPatchBatch
 from wsi_patching.writers.stream_writer_base import StreamWriterBase
 
 
-class NumpyMemoryWriter(StreamWriterBase):
+class NumpyStreamWriter(StreamWriterBase):
     """
-    Collects CollatedPatchBatch (assumed BCHW) and builds an in-memory NumPy dataset.
-    Eagerly copies to float32 NumPy arrays; no torch involved.
+    Returns the collated patches as numpy arrays in a streaming fashion.
     """
 
     def __init__(self, layout: Literal["NCHW", "NHWC"] = "NCHW", dtype: np.dtype = np.float32) -> None:
         super().__init__()
         self.layout = layout
         self.dtype = dtype
-
-        self._images_chunks: List[np.ndarray] = []
-        self._coords_chunks: List[np.ndarray] = []
-        self.meta = []
-
-        self.wsi_ids: List[str] = []
-
-        self.final_images, self.final_coords = None, None
 
     def stream(self, batch: CollatedPatchBatch) -> Iterable[Tuple[str, np.ndarray, np.ndarray, dict]]:
         self.log.info(f"Received batch from wsi: {batch.wsi_id} size: {len(batch.patches)}")
